requirements.txt
kafka-python==2.0.2
tenacity==8.2.2
pydantic==1.10.12
redis==4.5.5
prometheus-client==0.16.0
pytest==7.4.2
python-dotenv==1.0.0
croniter==1.3.15
